#632 - Move to new work package page folder structure
